refactor(predictions): share the sports schedule classifier
