+++ src/features/polymarket/screens/polymarket-sports-events-screen/PolymarketSportsEventsList.tsx
@@ -1,7 +1,6 @@
 import { memo, useCallback, useMemo, type RefObject } from 'react';
 import { StyleSheet, View, type NativeScrollEvent, type NativeSyntheticEvent, type ViewToken } from 'react-native';
 
-import { format } from 'date-fns';
 import { debounce } from 'lodash';
 import Animated from 'react-native-reanimated';
 import { useSafeAreaInsets } from 'react-native-safe-area-context';
@@ -14,10 +13,13 @@ import {
   PolymarketSportEventListItem,
 } from '@/features/polymarket/components/polymarket-sport-event-list-item/PolymarketSportEventListItem';
 import { DEFAULT_SPORTS_LEAGUE_KEY, NAVIGATOR_FOOTER_CLEARANCE, NAVIGATOR_FOOTER_HEIGHT } from '@/features/polymarket/constants';
-import { getLeague, getLeagueId, getLeagueSlugId, LEAGUE_LIST_ORDER } from '@/features/polymarket/leagues';
+import { getLeagueId } from '@/features/polymarket/leagues';
+import {
+  buildPolymarketSportsEventsListData,
+  type SportsListItem,
+} from '@/features/polymarket/screens/polymarket-sports-events-screen/buildPolymarketSportsEventsListData';
 import { usePolymarketSportsEventsStore } from '@/features/polymarket/stores/polymarketSportsEventsStore';
 import { type PolymarketEvent } from '@/features/polymarket/types/polymarket-event';
-import { getSportsEventsDayBoundaries } from '@/features/polymarket/utils/getSportsEventsDateRange';
 import { getSportsEventTokenIds } from '@/features/polymarket/utils/sportsEventBetData';
 import { useStableValue } from '@/hooks/useStableValue';
 import * as i18n from '@/languages';
@@ -34,38 +36,6 @@ type SportsEventsListProps = {
   onScroll?: (event: NativeSyntheticEvent<NativeScrollEvent>) => void;
 };
 
-type EventItem = {
-  type: 'event';
-  key: string;
-  event: PolymarketEvent;
-};
-
-type SectionHeaderItem = {
-  type: 'header';
-  key: string;
-  title: string;
-  isLive?: boolean;
-};
-
-type LeagueHeaderItem = {
-  type: 'league-header';
-  key: string;
-  title: string;
-  eventSlug: string;
-};
-
-type SectionSeparatorItem = {
-  type: 'separator';
-  key: string;
-};
-
-type LeagueSeparatorItem = {
-  type: 'league-separator';
-  key: string;
-};
-
-export type SportsListItem = EventItem | SectionHeaderItem | LeagueHeaderItem | SectionSeparatorItem | LeagueSeparatorItem;
-
 export const PolymarketSportsEventsList = memo(function PolymarketSportsEventsList({ listRef, onScroll }: SportsEventsListProps) {
   const safeAreaInsets = useSafeAreaInsets();
   const events = usePolymarketSportsEventsStore(state => state.getData() ?? EMPTY_EVENTS);
@@ -78,7 +48,10 @@ export const PolymarketSportsEventsList = memo(function PolymarketSportsEventsLi
     return events.filter(event => getLeagueId(event.slug) === selectedLeagueId);
   }, [events, showLeagueHeaders, selectedLeagueId]);
 
-  const listData = useMemo(() => buildListData(filteredEvents, showLeagueHeaders), [filteredEvents, showLeagueHeaders]);
+  const listData = useMemo(
+    () => buildPolymarketSportsEventsListData(filteredEvents, showLeagueHeaders),
+    [filteredEvents, showLeagueHeaders]
+  );
 
   const listStyles = useMemo(() => {
     const paddingBottom = safeAreaInsets.bottom + NAVIGATOR_FOOTER_HEIGHT + NAVIGATOR_FOOTER_CLEARANCE;
@@ -228,162 +201,6 @@ function extractEventTokenIds(viewTokens: Array<ViewToken<SportsListItem>>): str
     }
   }
   return Array.from(tokenIds);
-}
-
-function buildListData(events: PolymarketEvent[], showLeagueHeaders: boolean): SportsListItem[] {
-  if (!events.length) return [];
-
-  const { startOfToday, startOfTomorrow, startOfDayAfterTomorrow } = getSportsEventsDayBoundaries();
-  const startOfTodayMs = startOfToday.getTime();
-  const startOfTomorrowMs = startOfTomorrow.getTime();
-  const startOfDayAfterTomorrowMs = startOfDayAfterTomorrow.getTime();
-
-  const liveEvents: PolymarketEvent[] = [];
-  const todayEvents: PolymarketEvent[] = [];
-  const tomorrowEvents: PolymarketEvent[] = [];
-
-  for (const event of events) {
-    if (event.live && !event.ended) {
-      liveEvents.push(event);
-      continue;
-    }
-
-    const startTime = getTimestamp(event.startTime);
-    const endTime = getTimestamp(event.endDate);
-    const bucket = getTimeBucket({
-      timestamp: startTime ?? endTime ?? startOfTodayMs,
-      startOfTodayMs,
-      startOfTomorrowMs,
-      startOfDayAfterTomorrowMs,
-    });
-
-    if (bucket === 'today') {
-      todayEvents.push(event);
-      continue;
-    }
-    if (bucket === 'tomorrow') {
-      tomorrowEvents.push(event);
-      continue;
-    }
-
-    if (startTime != null && endTime != null && startTime !== endTime) {
-      const fallbackBucket = getTimeBucket({ timestamp: endTime, startOfTodayMs, startOfTomorrowMs, startOfDayAfterTomorrowMs });
-      if (fallbackBucket === 'today') {
-        todayEvents.push(event);
-      } else if (fallbackBucket === 'tomorrow') {
-        tomorrowEvents.push(event);
-      }
-    }
-  }
-
-  const items: SportsListItem[] = [];
-  const tomorrowLabel = format(startOfTomorrow, 'EEE, MMM d', { locale: i18n.getDateFnsLocale() });
-
-  const sections = [
-    { key: 'live', title: i18n.t(i18n.l.predictions.sports.live), events: liveEvents },
-    { key: 'today', title: i18n.t(i18n.l.time.today_caps), events: todayEvents },
-    { key: 'tomorrow', title: tomorrowLabel, events: tomorrowEvents },
-  ];
-
-  for (const section of sections) {
-    if (!section.events.length) continue;
-    if (items.length) {
-      items.push({ type: 'separator', key: `separator-${section.key}` });
-    }
-    pushSection({ items, title: section.title, events: section.events, sectionKey: section.key, showLeagueHeaders });
-  }
-
-  return items;
-}
-
-function pushSection({
-  items,
-  title,
-  events,
-  sectionKey,
-  showLeagueHeaders,
-}: {
-  items: SportsListItem[];
-  title: string;
-  events: PolymarketEvent[];
-  sectionKey: string;
-  showLeagueHeaders: boolean;
-}) {
-  if (!events.length) return;
-  items.push({ type: 'header', key: `header-${sectionKey}`, title, isLive: sectionKey === 'live' });
-  const leagueGroups = groupEventsByLeague(events);
-  for (let i = 0; i < leagueGroups.length; i++) {
-    const group = leagueGroups[i];
-    if (showLeagueHeaders) {
-      if (i > 0) {
-        items.push({ type: 'league-separator', key: `league-separator-${sectionKey}-${group.key}` });
-      }
-      items.push({ type: 'league-header', key: `league-${sectionKey}-${group.key}`, title: group.title, eventSlug: group.events[0].slug });
-    }
-    items.push(...buildEventItems(group.events, `${sectionKey}-${group.key}`));
-  }
-}
-
-function buildEventItems(events: PolymarketEvent[], sectionKey: string): EventItem[] {
-  return events.map((event, index) => ({
-    type: 'event',
-    key: `event-${sectionKey}-${event.id}-${index}`,
-    event,
-  }));
-}
-
-function groupEventsByLeague(events: PolymarketEvent[]) {
-  const groups = new Map<string, { key: string; title: string; events: PolymarketEvent[] }>();
-  for (const event of events) {
-    const leagueId = getLeagueId(event.slug);
-    const league = getLeague(event.slug);
-    const leagueSlugId = getLeagueSlugId(event.slug);
-    const key = leagueId ?? leagueSlugId ?? 'unknown';
-    const title = league?.name ?? (leagueSlugId ? leagueSlugId.toUpperCase() : i18n.t(i18n.l.predictions.bet_types.other));
-    const existing = groups.get(key) ?? { key, title, events: [] };
-    existing.events.push(event);
-    groups.set(key, existing);
-  }
-
-  // Sort events within each group by earliest start time
-  for (const group of groups.values()) {
-    group.events.sort((a, b) => {
-      const aTime = getTimestamp(a.startTime) ?? Infinity;
-      const bTime = getTimestamp(b.startTime) ?? Infinity;
-      return aTime - bTime;
-    });
-  }
-
-  return Array.from(groups.values()).sort((a, b) => {
-    const aIndex = LEAGUE_LIST_ORDER.indexOf(a.key as (typeof LEAGUE_LIST_ORDER)[number]);
-    const bIndex = LEAGUE_LIST_ORDER.indexOf(b.key as (typeof LEAGUE_LIST_ORDER)[number]);
-    if (aIndex === -1 && bIndex === -1) return 0;
-    if (aIndex === -1) return 1;
-    if (bIndex === -1) return -1;
-    return aIndex - bIndex;
-  });
-}
-
-function getTimestamp(value?: string): number | null {
-  if (!value) return null;
-  const timestamp = new Date(value).getTime();
-  return Number.isNaN(timestamp) ? null : timestamp;
-}
-
-function getTimeBucket({
-  timestamp,
-  startOfTodayMs,
-  startOfTomorrowMs,
-  startOfDayAfterTomorrowMs,
-}: {
-  timestamp: number;
-  startOfTodayMs: number;
-  startOfTomorrowMs: number;
-  startOfDayAfterTomorrowMs: number;
-}) {
-  if (timestamp >= startOfTodayMs && timestamp < startOfTomorrowMs) return 'today';
-  if (timestamp >= startOfTomorrowMs && timestamp < startOfDayAfterTomorrowMs) return 'tomorrow';
-  return null;
 }
 
 const styles = StyleSheet.create({
